feat: Implement Stock and Product Client Ports
- Added Stock model with fields for id, productId, enterpriseId, quantity, price, and status.
- Created IProductClientPort interface for fetching products by enterprise ID.
- Created IStockClientPort interface for buying stock.
- Configured RabbitMQ settings for message handling with error logging and acknowledgment.
- Implemented RabbitWeightedAverageConfig for managing exchanges and queues with dead letter and retry configurations.
- Developed KardexDtoRequest for handling purchase requests with validation.
- Implemented FactureListener for processing messages from RabbitMQ and handling events.
- Added enums for event types related to products and invoices.
- Created ProductClientAdapter and StockClientAdapter for remote service communication.
- Configured WebClient for making HTTP requests to product and stock services.
- Developed mappers for converting between domain models and DTOs.
- Added unit tests for Kardex model to ensure correct behavior for various operations.

+++ src/main/java/com/kardex/domain/model/Kardex.java
@@ -17,7 +17,7 @@ public class Kardex {
 
     private Long factCode;
 
-    private Long quantity;
+    private int quantity;
 
     private BigDecimal unitPrice;
 
@@ -25,7 +25,7 @@ public class Kardex {
 
     private MovementType type;
 
-    private Long balanceQuantity;
+    private int balanceQuantity;
 
     private BigDecimal balanceUnitPrice;
 
@@ -39,7 +39,7 @@ public class Kardex {
         this.date = ZonedDateTime.now(ZoneId.of("America/Bogota"));
     }
 
-    public void addPurchase(Long lastQuantity, BigDecimal lastTotalBalance) {
+    public void addPurchase(int lastQuantity, BigDecimal lastTotalBalance) {
         // 1. Calcula la nueva cantidad total en el balance
         this.balanceQuantity = lastQuantity + this.quantity;
 
@@ -64,7 +64,7 @@ public class Kardex {
         this.updateDetailIfNotNull();
     }
 
-    public void addSale(Long lastQuantity, BigDecimal lastUnitPrice, BigDecimal lastTotalBalance) {
+    public void addSale(int lastQuantity, BigDecimal lastUnitPrice, BigDecimal lastTotalBalance) {
         // 1. Calcula la nueva cantidad total en el balance
         this.balanceQuantity = lastQuantity - this.quantity;
         
@@ -89,7 +89,7 @@ public class Kardex {
         this.updateDetailIfNotNull();   
     }
 
-    public void returnOnSale(Long lastQuantity, BigDecimal lastUnitPrice, BigDecimal lastTotalBalance) {
+    public void returnOnSale(int lastQuantity, BigDecimal lastUnitPrice, BigDecimal lastTotalBalance) {
         // 1. Calcula la nueva cantidad total en el balance
         this.balanceQuantity = lastQuantity + this.quantity;
         
@@ -107,7 +107,7 @@ public class Kardex {
         this.updateDetailIfNotNull();      
     }
 
-    public void returnOnPurchase(Long lastQuantity, BigDecimal lastTotalBalance) {
+    public void returnOnPurchase(int lastQuantity, BigDecimal lastTotalBalance) {
         // 1. Calcula la nueva cantidad total en el balance
         this.balanceQuantity = lastQuantity - this.quantity;
         
@@ -137,7 +137,7 @@ public class Kardex {
     public void resetBalancesIfZero(){
             this.totalBalance = BigDecimal.ZERO;
             this.balanceUnitPrice = BigDecimal.ZERO;
-            this.balanceQuantity = 0L;     
+            this.balanceQuantity = 0;     
     }
 
     public void updateDetailIfNotNull() {
